feat: Implemented BCSInit function for value initialization after decoding; Splitted tests into multiple files
